refactored PlayerRanking #12
