+++ src/styles/views/Game/PlayerRanking.scss
@@ -1,0 +1,74 @@
+@import "styles/theme";
+
+.ranking-container {
+  height: 90%;
+  width: 90%;
+  border-radius: $box-radius;
+  box-shadow: $box-shadow;
+  background-color: $background-color;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+
+.ranking-table {
+  height: 90%;
+  width: 90%;
+  //  background-color: gray;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  flex-direction: column;
+}
+
+.ranking-row {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 90%;
+  height: 10%;
+  margin-bottom: 4%;
+}
+
+.ranking-element {
+  border-radius: $box-radius;
+  background-color: $list-color;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+
+  &.rank {
+    width: 10%;
+    font-weight: bold;
+  }
+  &.name {
+    width: 50%;
+  }
+  &.role {
+    width: 10%;
+  }
+  &.points {
+    width: 20%;
+  }
+}
+
+.ranking-header {
+  display: flex;
+  flex-direction: row;
+
+  & h1 {
+    padding-left: 50px;
+    font-size: 250%;
+    font-weight: 700;
+    align-items: center;
+    color: $accent-color;
+  }
+
+  & img {
+    height: auto;
+    width: auto;
+    padding-left: 40px;
+  }
+}
